+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+ - Sample Editor: Tweaked resampler lowpass-filter settings to preserve more treble.
+
+### Added
+ - Sample Editor: Preferences menu for low-pass filter and dither.
 
 ## [1.11.4] - 2021-06-01
 ### Fixed